refactor: refactor function in utils
