+++ omnisafe/algorithms/on_policy/second_order/cpo.py
@@ -22,8 +22,8 @@ import torch
 from omnisafe.algorithms import registry
 from omnisafe.algorithms.on_policy.base.trpo import TRPO
 from omnisafe.utils import distributed
+from omnisafe.utils.math import conjugate_gradients
 from omnisafe.utils.tools import (
-    conjugate_gradients,
     get_flat_gradients_from,
     get_flat_params_from,
     set_param_values_to_model,
